refactor: react query hook name 변경
