+++ src/Layout/ComponentLayout.tsx
@@ -1,13 +1,12 @@
-import React, { useEffect } from 'react';
-import { useCookies } from 'react-cookie';
-import { useGetUserToken } from '../api/hooks/useGetUserData';
-import Alert from '../components/common/Alert';
-import Modal from '../components/common/modal';
-import SideBar from '../components/common/SideBar';
-import Navigation from '../components/common/Navigation';
 import { AnimatePresence } from 'framer-motion';
-import GoogleLoader from '../components/common/GoogleLoader';
+import React from 'react';
 import { useRecoilState } from 'recoil';
+import { useGetMyToken } from '../api/hooks/useGetMyData';
+import Alert from '../components/common/Alert';
+import GoogleLoader from '../components/common/GoogleLoader';
+import Modal from '../components/common/modal';
+import Navigation from '../components/common/Navigation';
+import SideBar from '../components/common/SideBar';
 import { loaderState } from '../store/loader';
 import GlobalStyles from '../styles/globalStyles';
 import { NavigationBlock } from '../styles/layouts';
@@ -16,7 +15,7 @@ const ComponentLayout: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
   const [loader] = useRecoilState(loaderState);
-  useGetUserToken();
+  useGetMyToken();
   return (
     <div>
       <Alert />
